Commit Inicial, se agrega logo marca pais

+++ concrete/src/Page/Type/Type.php
@@ -153,7 +153,7 @@ class Type extends Object implements \Concrete\Core\Permission\ObjectInterface
         }
     }
 
-    public function publish(Page $c)
+    public function publish(Page $c, $pkr = null)
     {
         $this->stripEmptyPageTypeComposerControls($c);
         $parent = Page::getByID($c->getPageDraftTargetParentPageID());
@@ -173,13 +173,14 @@ class Type extends Object implements \Concrete\Core\Permission\ObjectInterface
         }
 
         $u = new User();
-        $v = CollectionVersion::get($c, 'RECENT');
-        $pkr = new ApprovePagePageWorkflowRequest();
-        $pkr->setRequestedPage($c);
-        $pkr->setRequestedVersionID($v->getVersionID());
-        $pkr->setRequesterUserID($u->getUserID());
+        if (!is_object($pkr)) {
+            $v = CollectionVersion::get($c, 'RECENT');
+            $pkr = new ApprovePagePageWorkflowRequest();
+            $pkr->setRequestedPage($c);
+            $pkr->setRequestedVersionID($v->getVersionID());
+            $pkr->setRequesterUserID($u->getUserID());
+        }
         $pkr->trigger();
-
         $u->unloadCollectionEdit($c);
         CacheLocal::flush();
 
@@ -218,14 +219,16 @@ class Type extends Object implements \Concrete\Core\Permission\ObjectInterface
         }
 
         // remove all but the most recent X drafts.
-        $vl = new VersionList($c);
-        $vl->setItemsPerPage(-1);
-        // this will ensure that we only ever keep X versions.
-        $vArray = $vl->getPage();
-        if (count($vArray) > $this->ptDraftVersionsToSave) {
-            for ($i = $this->ptDraftVersionsToSave; $i < count($vArray); $i++) {
-                $v = $vArray[$i];
-                @$v->delete();
+        if ($c->isPageDraft()) {
+            $vl = new VersionList($c);
+            $vl->setItemsPerPage(-1);
+            // this will ensure that we only ever keep X versions.
+            $vArray = $vl->getPage();
+            if (count($vArray) > $this->ptDraftVersionsToSave) {
+                for ($i = $this->ptDraftVersionsToSave; $i < count($vArray); $i++) {
+                    $v = $vArray[$i];
+                    @$v->delete();
+                }
             }
         }
 
@@ -1120,9 +1123,8 @@ class Type extends Object implements \Concrete\Core\Permission\ObjectInterface
         $db = Loader::db();
         $ptID = $this->getPageTypeID();
         $parent = Page::getByPath(Config::get('concrete.paths.drafts'));
-        $data = array('cvIsApproved' => 0);
+        $data = array('cvIsApproved' => 0, 'cIsActive' => false, 'cAcquireComposerOutputControls' => true);
         $p = $parent->add($this, $data, $pt);
-        $p->deactivate();
 
         // now we setup in the initial configurated page target
         $target = $this->getPageTypePublishTargetObject();
